% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heteroclitic.R
\name{heteroclitic}
\alias{heteroclitic}
\title{Generates single amino-acid variant of a peptide.}
\usage{
heteroclitic(peptide, variant = "all", case = "upper")
}
\arguments{
\item{variant}{Default status is to return all possible variants}

\item{case}{Default is to output upper case unless "lower" is applied}
}
\description{
This function mutates a peptide sequence and outputs single amino-acid variants of that peptide.
}
\examples{
heteroclitic("LLLALLPPGA")

heteroclitic("LLLALLPPGA", variant = "4R", case = "lower")
}
