% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window.R
\name{sliding_window}
\alias{sliding_window}
\title{Outputs all peptides consituting longer protein sequence}
\usage{
sliding_window(prot.sequence, pep.len = 9)
}
\arguments{
\item{pep.len}{Peptide length default to 9}
}
\description{
This function generates a list of peptides of desired lengths that compose a longer protein in table format, along with position and length infomration.
}
\examples{
prot.sequence = "ETFNTPAMYVAIQAVLSLYASGRTTGIVMDSGDG"
sliding_window(prot.sequence, 8:12)
}
