% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_peptides.R
\name{snv_peptides}
\alias{snv_peptides}
\title{Outputs all peptides containing a specific point mutation}
\usage{
snv_peptides(prot_sequence = NULL, mutation = NULL, pep_length = 9)
}
\arguments{
\item{pep_length}{Peptide length default to 9}
}
\description{
This function generates a list of peptides of desired lengths that incorporate a specific point mutation, along with position and length information, as well as WT equivalent peptide.
}
\examples{
protein = "ETFNTPAMYVAIQAVLSLYASGRTTGIVMDSGDG"
mut = "V15L"
peptide.sizes = c(9,10)
snv_peptides(prot_sequence = protein, mutation = "V15L", pep_length = peptide.sizes)
}
